+++ packages/plugin-swr/src/generators/queryGenerator.tsx
@@ -33,7 +33,7 @@ export const queryGenerator = createReactGenerator<PluginSwr>({
     }
 
     const client = {
-      name: getName(operation, { type: 'function', pluginKey: [pluginClientName] }),
+      name: getName(operation, { type: 'function' }),
     }
 
     const queryOptions = {